+++ extremecrawler/exceptions/__init__.py
@@ -1,0 +1,2 @@
+class NotHtmlError(BaseException):
+    pass
