--- page_lister/__init__.py
@@ -1,1 +1,0 @@
-from .main import get_links
