+++ tests/generic_test.py
@@ -3,9 +3,9 @@ import unittest
 class GenericTests(unittest.TestCase):
 
     def test_import(self):
-        import page_lister
+        from extremecrawler import ExtremeCrawler
 
 
     def test_accessible(self):
-        import page_lister
-        page_lister.get_links
+        from extremecrawler import ExtremeCrawler
+
